feat: PATCH /api/room/id
